+++ src/main/java/roomescape/dto/request/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto.request;
+package roomescape.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import java.time.LocalTime;
