+++ src/main/java/roomescape/dto/request/ThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto.request;
+package roomescape.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
