+++ src/main/java/roomescape/dto/response/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto.response;
+package roomescape.dto.response;
 
 import java.time.LocalDate;
 import roomescape.domain.Reservation;
