+++ src/main/java/roomescape/dto/response/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto.response;
+package roomescape.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
