+++ src/test/java/roomescape/controller/BaseControllerTest.java
@@ -7,7 +7,7 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.annotation.DirtiesContext.ClassMode;
-import roomescape.service.config.TestConfig;
+import roomescape.config.TestConfig;
 
 @SpringBootTest(
         classes = TestConfig.class,
